Extract the HazmatCard into a dedicated component code due to the reuse
